add arch=native and change the size of Square from 8 bit to 32 bit.

+++ c++_version/main.cpp
@@ -2,7 +2,7 @@
 #include <chrono>
 #include <cstdint>
 
-enum Square : int8_t {};
+enum Square : int32_t {};
 
 struct Bitboard {
     uint64_t value(size_t i) const { return v[i]; }
